Move integration code to separate compilation unit
USING_MAIN is now defined in CMakeLists
