+++ sim/main_integration.h
@@ -1,0 +1,6 @@
+
+extern void main_integration_feed_simstate(simstate_t* simstate);
+extern void main_integration_command(int cmd);
+extern void main_integration_simLoop();
+extern void main_integration_restart();
+extern void main_integration_init();
